only TracerProvider.Dispose will call processor.Shutdown
