+++ src/OpenTelemetry/Trace/BatchExportActivityProcessor.cs
@@ -25,9 +25,8 @@ namespace OpenTelemetry.Trace
     /// <summary>
     /// Implements processor that batches activities before calling exporter.
     /// </summary>
-    public class BatchExportActivityProcessor : ActivityProcessor
+    public class BatchExportActivityProcessor : BaseExportActivityProcessor
     {
-        private readonly ActivityExporter exporter;
         private readonly CircularBuffer<Activity> circularBuffer;
         private readonly int scheduledDelayMilliseconds;
         private readonly int exporterTimeoutMilliseconds;
@@ -37,7 +36,6 @@ namespace OpenTelemetry.Trace
         private readonly ManualResetEvent dataExportedNotification = new ManualResetEvent(false);
         private readonly ManualResetEvent shutdownTrigger = new ManualResetEvent(false);
         private long shutdownDrainTarget = long.MaxValue;
-        private bool disposed;
         private long droppedCount = 0;
 
         /// <summary>
@@ -54,6 +52,7 @@ namespace OpenTelemetry.Trace
             int scheduledDelayMilliseconds = 5000,
             int exporterTimeoutMilliseconds = 30000,
             int maxExportBatchSize = 512)
+            : base(exporter)
         {
             if (maxQueueSize <= 0)
             {
@@ -75,7 +74,6 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentOutOfRangeException(nameof(exporterTimeoutMilliseconds));
             }
 
-            this.exporter = exporter ?? throw new ArgumentNullException(nameof(exporter));
             this.circularBuffer = new CircularBuffer<Activity>(maxQueueSize);
             this.scheduledDelayMilliseconds = scheduledDelayMilliseconds;
             this.exporterTimeoutMilliseconds = exporterTimeoutMilliseconds;
@@ -184,29 +182,6 @@ namespace OpenTelemetry.Trace
             if (timeoutMilliseconds != 0)
             {
                 this.exporterThread.Join(timeoutMilliseconds);
-            }
-        }
-
-        /// <inheritdoc/>
-        protected override void Dispose(bool disposing)
-        {
-            base.Dispose(disposing);
-
-            if (disposing && !this.disposed)
-            {
-                // TODO: Dispose/Shutdown flow needs to be redesigned, currently it is convoluted.
-                this.Shutdown(this.exporterTimeoutMilliseconds);
-
-                try
-                {
-                    this.exporter.Dispose();
-                }
-                catch (Exception ex)
-                {
-                    OpenTelemetrySdkEventSource.Log.SpanProcessorException(nameof(this.Dispose), ex);
-                }
-
-                this.disposed = true;
             }
         }
